pyproject.toml
[project]
name = "knockoff-muzero"
version = "0.1.0"
description = "A from-scratch MuZero knockoff for IT-3105"
requires-python = ">=3.10"
dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "equinox>=0.11",
    "optax>=0.1.7",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pygame>=2.5",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["muzero*"]
